requirements.txt
python-dotenv
sqlalchemy
oracledb


